refactor(data-warehouse): route data_warehouse's ws model/type reads through the facade
Routes data_warehouse's own non-presentation imports of warehouse_sources models/types (logic, temporal activities, tasks) through ws.facade (per-symbol: model classes -> facade.models, ClickHouse mapping tables -> facade.hogql, enums -> facade.types). Shrinks the largest chunk of the ws legacy-leak. Adds the few missing symbols to ws.facade.models (SSHTunnel, acreate/asave_datawarehousetable, mysql/postgres_column_to_dwh_column). Setup-path modules (direct_mysql/postgres, the dw models) keep direct ws.models imports — routing them through the eager facade.models re-enters mid-load (ws.models.table -> dw.facade.sources -> direct_mysql), a justified setup-path coupling.
