+++ products/data_warehouse/backend/logic/external_data_source/jobs.py
@@ -10,8 +10,7 @@ from products.data_warehouse.backend.tasks import (
     EXTERNAL_DATA_FAILURE_DIGEST_SCHEDULED_COUNTER,
     send_external_data_failure_digest_task,
 )
-from products.warehouse_sources.backend.models.external_data_job import ExternalDataJob
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
+from products.warehouse_sources.backend.facade.models import ExternalDataJob, ExternalDataSchema
 from products.warehouse_sources.backend.temporal.data_imports.metrics import (
     TERMINAL_JOB_STATUSES,
     emit_data_import_app_metrics,
