+++ products/data_warehouse/backend/logic/external_data_source/webhooks.py
@@ -7,7 +7,7 @@ from posthog.models import Team
 
 from products.cdp.backend.models.hog_function_template import HogFunctionTemplate
 from products.cdp.backend.models.hog_functions.hog_function import HogFunction
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
+from products.warehouse_sources.backend.facade.models import ExternalDataSchema
 from products.warehouse_sources.backend.temporal.data_imports.sources.common.base import (
     WebhookCreationResult,
     WebhookDeletionResult,
