+++ products/data_warehouse/backend/management/commands/backfill_cdc_extraction_schedules.py
@@ -35,8 +35,7 @@ from django.core.management.base import BaseCommand
 import structlog
 
 from products.data_warehouse.backend.logic.data_load.service import bulk_sync_cdc_extraction_schedules, cdc_min_interval
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
+from products.warehouse_sources.backend.facade.models import ExternalDataSchema, ExternalDataSource
 
 logger = structlog.get_logger(__name__)
