+++ products/data_warehouse/backend/management/commands/backfill_discovery_schedules.py
@@ -33,9 +33,9 @@ from django.core.management.base import BaseCommand
 import structlog
 
 from products.data_warehouse.backend.logic.data_load.service import bulk_sync_discover_schemas_schedules
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
+from products.warehouse_sources.backend.facade.models import ExternalDataSource
+from products.warehouse_sources.backend.facade.types import ExternalDataSourceType
 from products.warehouse_sources.backend.temporal.data_imports.sources import SourceRegistry
-from products.warehouse_sources.backend.types import ExternalDataSourceType
 
 logger = structlog.get_logger(__name__)
