+++ products/data_warehouse/backend/management/test/test_backfill_discovery_schedules.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 
 from django.core.management import call_command
 
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
+from products.warehouse_sources.backend.facade.models import ExternalDataSource
 
 pytestmark = pytest.mark.django_db
