+++ products/data_warehouse/backend/management/test/test_update_data_import_schedules.py
@@ -16,8 +16,7 @@ from posthog.temporal.common.schedule import describe_schedule, update_schedule
 
 from products.data_warehouse.backend.logic.data_load.service import sync_external_data_job_workflow
 from products.data_warehouse.backend.management.commands.update_data_import_schedules import _get_external_data_schemas
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
+from products.warehouse_sources.backend.facade.models import ExternalDataSchema, ExternalDataSource
 
 pytestmark = [pytest.mark.django_db]
