+++ products/data_warehouse/backend/postgres_warehouse_migration.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 from products.data_warehouse.backend.postgres_helpers import rename_postgres_schemas_to_match_source_schemas
 from products.data_warehouse.backend.sql_warehouse_migration import apply_on_refresh
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
+from products.warehouse_sources.backend.facade.models import ExternalDataSource
 from products.warehouse_sources.backend.temporal.data_imports.sources.common.schema import SourceSchema
