+++ products/data_warehouse/backend/sql_warehouse_migration.py
@@ -9,7 +9,8 @@ from __future__ import annotations
 
 from typing import Any
 
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
+from products.warehouse_sources.backend.facade.models import ExternalDataSource
+from products.warehouse_sources.backend.facade.types import ExternalDataSourceType
 from products.warehouse_sources.backend.temporal.data_imports.naming_convention import NamingConvention
 from products.warehouse_sources.backend.temporal.data_imports.sources.common.registry import SourceRegistry
 from products.warehouse_sources.backend.temporal.data_imports.sources.common.sql.base import SQLSource
@@ -17,7 +18,6 @@ from products.warehouse_sources.backend.temporal.data_imports.sources.common.sql
     fill_missing_from_dotted_name,
     normalize_namespace,
 )
-from products.warehouse_sources.backend.types import ExternalDataSourceType
 
 
 def _source_has_optional_schema_field(source: Any) -> bool:
@@ -103,7 +103,7 @@ def apply_on_schema_clear(source: ExternalDataSource, old_schema: str) -> None:
     """Pin legacy rows to the OLD schema before the next refresh sees `default_schema=None` and
     misroutes them to `"public"`.
     """
-    from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
+    from products.warehouse_sources.backend.facade.models import ExternalDataSchema
 
     rows = list(ExternalDataSchema.objects.filter(team_id=source.team_id, source_id=source.id, deleted=False))
     rows_by_name = {row.name: row for row in rows}
@@ -163,7 +163,7 @@ def apply_on_refresh(*, source: ExternalDataSource, team_id: int) -> dict[str, s
     falls back to `"public"` when `job_inputs.schema` is blank, missing the actual schema.
     Returns `{old_name: new_name}` for callers feeding `sync_old_schemas_with_new_schemas`.
     """
-    from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
+    from products.warehouse_sources.backend.facade.models import ExternalDataSchema
 
     rows = list(
         ExternalDataSchema.objects.filter(team_id=team_id, source_id=source.id, deleted=False).select_related("table")
