+++ products/data_warehouse/backend/test/test_repair_qualified_schema_duplicates.py
@@ -6,10 +6,8 @@ from unittest.mock import patch
 
 from django.core.management import call_command
 
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
-from products.warehouse_sources.backend.models.table import DataWarehouseTable
-from products.warehouse_sources.backend.types import ExternalDataSourceType
+from products.warehouse_sources.backend.facade.models import DataWarehouseTable, ExternalDataSchema, ExternalDataSource
+from products.warehouse_sources.backend.facade.types import ExternalDataSourceType
 
 # Patched at the defining module: update_should_sync imports these function-locally now.
 _TEMPORAL = "products.data_warehouse.backend.logic.data_load.service"
