+++ products/data_warehouse/backend/test/test_sync_status.py
@@ -6,11 +6,13 @@ from unittest.mock import MagicMock
 from parameterized import parameterized
 
 from products.data_warehouse.backend.sync_status import get_warehouse_sync_warnings
-from products.warehouse_sources.backend.models.credential import DataWarehouseCredential
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
-from products.warehouse_sources.backend.models.table import DataWarehouseTable
-from products.warehouse_sources.backend.types import ExternalDataSourceType
+from products.warehouse_sources.backend.facade.models import (
+    DataWarehouseCredential,
+    DataWarehouseTable,
+    ExternalDataSchema,
+    ExternalDataSource,
+)
+from products.warehouse_sources.backend.facade.types import ExternalDataSourceType
 
 
 class TestWarehouseSyncWarnings(BaseTest):
