+++ products/data_warehouse/backend/tests/api/test_lineage.py
@@ -4,7 +4,7 @@ from posthog.test.db_context_capturing import capture_db_queries
 
 from products.data_modeling.backend.models.datawarehouse_saved_query import DataWarehouseSavedQuery
 from products.data_warehouse.backend.presentation.views.lineage import topological_sort
-from products.warehouse_sources.backend.models.table import DataWarehouseTable
+from products.warehouse_sources.backend.facade.models import DataWarehouseTable
 
 
 class TestLineage(APIBaseTest):
