+++ products/data_warehouse/backend/tests/api/test_postgres_warehouse_migration.py
@@ -13,9 +13,7 @@ from unittest.mock import Mock, patch
 
 from rest_framework import status
 
-from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
-from products.warehouse_sources.backend.models.table import DataWarehouseTable
+from products.warehouse_sources.backend.facade.models import DataWarehouseTable, ExternalDataSchema, ExternalDataSource
 from products.warehouse_sources.backend.temporal.data_imports.sources.common.schema import SourceSchema
