+++ products/data_warehouse/backend/tests/api/test_table.py
@@ -13,8 +13,7 @@ from parameterized import parameterized
 
 from products.data_warehouse.backend.direct_postgres import DIRECT_POSTGRES_URL_PATTERN
 from products.data_warehouse.backend.presentation.views.table import SimpleTableSerializer
-from products.warehouse_sources.backend.models.external_data_source import ExternalDataSource
-from products.warehouse_sources.backend.models.table import DataWarehouseTable
+from products.warehouse_sources.backend.facade.models import DataWarehouseTable, ExternalDataSource
 
 
 class TestTable(APIBaseTest):
@@ -646,7 +645,7 @@ class TestTable(APIBaseTest):
         assert table.url_pattern == "https://your-org.s3.amazonaws.com/bucket/whatever.pqt"
 
     def test_update_table_credential_blank_access_key(self):
-        from products.warehouse_sources.backend.models.credential import DataWarehouseCredential
+        from products.warehouse_sources.backend.facade.models import DataWarehouseCredential
 
         credential = DataWarehouseCredential.objects.create(
             team=self.team, access_key="original_key", access_secret="original_secret"
@@ -669,7 +668,7 @@ class TestTable(APIBaseTest):
         assert credential.access_key == "original_key"
 
     def test_update_table_credential_blank_access_secret(self):
-        from products.warehouse_sources.backend.models.credential import DataWarehouseCredential
+        from products.warehouse_sources.backend.facade.models import DataWarehouseCredential
 
         credential = DataWarehouseCredential.objects.create(
             team=self.team, access_key="original_key", access_secret="original_secret"
@@ -692,7 +691,7 @@ class TestTable(APIBaseTest):
         assert credential.access_secret == "original_secret"
 
     def test_update_table_credential_null_field_values_rejected(self):
-        from products.warehouse_sources.backend.models.credential import DataWarehouseCredential
+        from products.warehouse_sources.backend.facade.models import DataWarehouseCredential
 
         credential = DataWarehouseCredential.objects.create(
             team=self.team, access_key="original_key", access_secret="original_secret"
@@ -717,7 +716,7 @@ class TestTable(APIBaseTest):
         assert credential.access_secret == "original_secret"
 
     def test_update_table_credential_null_rejected(self):
-        from products.warehouse_sources.backend.models.credential import DataWarehouseCredential
+        from products.warehouse_sources.backend.facade.models import DataWarehouseCredential
 
         credential = DataWarehouseCredential.objects.create(
             team=self.team, access_key="original_key", access_secret="original_secret"
