+++ products/data_warehouse/backend/tests/api/test_table_access_control.py
@@ -7,7 +7,7 @@ from posthog.models.organization import OrganizationMembership
 from posthog.models.user import User
 
 from products.data_warehouse.backend.tests.api._access_control_base import WarehouseAccessControlTestMixin
-from products.warehouse_sources.backend.models.table import DataWarehouseTable
+from products.warehouse_sources.backend.facade.models import DataWarehouseTable
 
 try:
     from ee.models.rbac.access_control import AccessControl
