+++ products/warehouse_sources/backend/facade/models.py
@@ -32,9 +32,17 @@ from products.warehouse_sources.backend.models.external_data_source import (
     get_direct_external_data_source_for_connection,
 )
 from products.warehouse_sources.backend.models.pending_source_credential import PendingSourceCredential
-from products.warehouse_sources.backend.models.table import DataWarehouseTable, DataWarehouseTableColumns
+from products.warehouse_sources.backend.models.ssh_tunnel import SSHTunnel
+from products.warehouse_sources.backend.models.table import (
+    DataWarehouseTable,
+    DataWarehouseTableColumns,
+    acreate_datawarehousetable,
+    asave_datawarehousetable,
+)
 from products.warehouse_sources.backend.models.util import (
+    mysql_column_to_dwh_column,
     mysql_columns_to_dwh_columns,
+    postgres_column_to_dwh_column,
     postgres_columns_to_dwh_columns,
     validate_source_prefix,
     validate_warehouse_table_url_pattern,
@@ -48,12 +56,17 @@ __all__ = [
     "ExternalDataSchema",
     "ExternalDataSource",
     "PendingSourceCredential",
+    "SSHTunnel",
     "WarehouseColumnAnnotation",
+    "acreate_datawarehousetable",
+    "asave_datawarehousetable",
     "get_all_schemas_for_source_id",
     "get_direct_external_data_source_for_connection",
     "get_latest_run_if_exists",
     "get_or_create_datawarehouse_credential",
+    "mysql_column_to_dwh_column",
     "mysql_columns_to_dwh_columns",
+    "postgres_column_to_dwh_column",
     "postgres_columns_to_dwh_columns",
     "sync_frequency_interval_to_sync_frequency",
     "sync_frequency_to_sync_frequency_interval",
